move selflink filter to cache
